Implement two-phase rendering
The first phase is only for handling events while the second is for actually building the DOM.
This way, we can avoid building the DOM in the first phase when it won't be used.

This also paves the way for providing a "request update" function to the user which only runs the second phase.